+++ lib/plurimath/xml_engine/oga/comment.rb
@@ -1,0 +1,8 @@
+module Plurimath
+  module XMLEngine
+    class Oga
+      class Comment < ::Oga::XML::Comment
+      end
+    end
+  end
+end
