+++ lib/plurimath/xml_engine/oga/document.rb
@@ -1,0 +1,10 @@
+require "plurimath/xml_engine/oga/node"
+
+module Plurimath
+  module XMLEngine
+    class Oga
+      class Document < Node
+      end
+    end
+  end
+end
