+++ lib/plurimath/xml_engine/oga/element.rb
@@ -1,0 +1,8 @@
+module Plurimath
+  module XMLEngine
+    class Oga
+      class Element < ::Oga::XML::Element
+      end
+    end
+  end
+end
