+++ lib/plurimath/xml_engine/ox/comment.rb
@@ -1,0 +1,13 @@
+require "ox/node"
+
+module Plurimath
+  module XMLEngine
+    class Ox
+      class Comment < ::Ox::Node
+        def comment?
+          true
+        end
+      end
+    end
+  end
+end
