+++ lib/plurimath/xml_engine/ox/node.rb
@@ -1,0 +1,13 @@
+require "ox/node"
+
+module Plurimath
+  module XMLEngine
+    class Ox
+      class Node < ::Ox::Node
+        def comment?
+          false
+        end
+      end
+    end
+  end
+end
